добавила точку по центру

+++ squares/medium.py
@@ -3,6 +3,6 @@ def medium(x,y,h,w):
     center_y = int(y+h/2)
     return center_x, center_y
 def second_square(x,y,h,w):
-    x2 = int(x - w/2)
-    y2 = int(y - h/2)
+    x2 = x - w/2
+    y2 = y - h/2
     return x2,y2
